requirements.txt
PyQt5==5.15.6
tensorflow==2.8.0
pillow==8.4.0
numpy==1.21.4
scipy==1.7.3
matplotlib==3.5.1
opencv-python==4.5.5.64
pyinstaller==4.5.1
